# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: Yoobe V3 API - Sistema de Orçamentos e Replicação
  description: |
    API para gerenciamento de orçamentos e replicação de produtos base.
    
    ## Fluxo de Orçamento → Aprovação → Replicação
    
    1. **Gestor** cria orçamento com produtos desejados
    2. **Admin Global** aprova ou rejeita o orçamento
    3. **Gestor** pode replicar produtos apenas após aprovação
    4. **Gestor** gerencia produtos replicados (ativar/inativar, editar)
    
    ## Autenticação
    
    Todas as rotas requerem autenticação via Bearer Token.
    - Role `admin`: Acesso total ao sistema
    - Role `manager`: Acesso limitado às funcionalidades de gestor
    - Role `user`: Acesso apenas a funcionalidades básicas
    
  version: 2.1.0
  contact:
    name: Yoobe Team
    email: support@yoobe.co
servers:
  - url: http://localhost:3000
    description: Servidor de Desenvolvimento
  - url: https://api.yoobe.co
    description: Servidor de Produção
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Budget:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID único do orçamento
        title:
          type: string
          description: Título do orçamento
        description:
          type: string
          description: Descrição detalhada
        status:
          type: string
          enum: [pending, approved, rejected]
          description: Status atual do orçamento
        total_amount:
          type: number
          format: float
          description: Valor total do orçamento
        company_id:
          type: string
          format: uuid
          description: ID da empresa do gestor
        manager_id:
          type: string
          format: uuid
          description: ID do gestor
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        reviewed_at:
          type: string
          format: date-time
          nullable: true
        reviewed_by:
          type: string
          format: uuid
          nullable: true
        admin_notes:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/BudgetItem'
    BudgetItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        budget_id:
          type: string
          format: uuid
        base_product_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
        custom_price:
          type: number
          format: float
        notes:
          type: string
          nullable: true
        base_product:
          $ref: '#/components/schemas/BaseProduct'
    BaseProduct:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        base_price:
          type: number
          format: float
        base_points_cost:
          type: integer
        image_url:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        category_id:
          type: string
          format: uuid
          nullable: true
    CompanyProduct:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        points_cost:
          type: integer
        stock_quantity:
          type: integer
        is_active:
          type: boolean
        company_id:
          type: string
          format: uuid
        base_product_id:
          type: string
          format: uuid
        budget_id:
          type: string
          format: uuid
        approved_at:
          type: string
          format: date-time
        approved_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string
          description: Mensagem de erro
        code:
          type: string
          description: Código de erro (opcional)
        details:
          type: object
          description: Detalhes adicionais do erro
paths:
...
  /api/gestor/orcamentos:
    get:
      summary: Listar orçamentos do gestor
      description: Retorna todos os orçamentos criados pelo gestor autenticado
      tags: [Gestor - Orçamentos]
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
          description: Filtrar por status
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Número da página
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
          description: Itens por página
      responses:
        '200':
          description: Lista de orçamentos
          content:
            application/json:
              schema:
                type: object
                properties:
                  budgets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Budget'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso negado - Role insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Criar novo orçamento
      description: Cria um novo orçamento com produtos selecionados
      tags: [Gestor - Orçamentos]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - items
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
                description:
                  type: string
                  maxLength: 1000
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required:
                      - base_product_id
                      - quantity
                    properties:
                      base_product_id:
                        type: string
                        format: uuid
                      quantity:
                        type: integer
                        minimum: 1
                      custom_price:
                        type: number
                        format: float
                        minimum: 0
                      notes:
                        type: string
                        maxLength: 500
      responses:
        '201':
          description: Orçamento criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  budget:
                    $ref: '#/components/schemas/Budget'
                  message:
                    type: string
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Não autorizado
        '403':
          description: Acesso negado - Role insuficiente

  # ========================================
  # GESTOR - PRODUTOS BASE
  # ========================================
...
  /api/gestor/produtos:
    get:
      summary: Listar produtos replicados
      description: Retorna produtos replicados da loja do gestor
      tags: [Gestor - Produtos Replicados]
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
          description: Filtrar por status ativo/inativo
        - name: search
          in: query
          schema:
            type: string
          description: Termo de busca
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Lista de produtos replicados
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/CompanyProduct'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer
...